refactor: simplify and modernize close button
Replaces the custom close.pdf bitmap with the xmark SF Symbol, renames
ToolbarCloseButton to CloseButton, adds an accessibility label, and
normalizes all 12 call sites to ToolbarItem(placement: .topBarTrailing).
