+++ Flipcash/Core/Screens/Main/Currency Swap/CurrencySellAmountScreen.swift
@@ -64,8 +64,8 @@ struct CurrencySellAmountScreen: View {
                 }
             }
             .toolbar {
-                ToolbarCloseButton {
-                    dismissAction()
+                ToolbarItem(placement: .topBarTrailing) {
+                    CloseButton { dismissAction() }
                 }
             }
             .dialog(item: $viewModel.dialogItem)
